class="fold" for issue #271
a respec post process function has been added to turn

section class="fold
h3 id=whatever

into an html details element.

this has been used in the operator dictionary section to add two folded views
and in one example in chapter 3  (3.3.2.3 examples) (search for FOLDING)

Currently the <h?> styling is lost and the section self link is lost
but the number and table of contents entry is mainained.

You _have_ to have an id on the h? first child of secion otherwise
I and rspec disagree on the id to use in the table of contents.

the javascript isn't that optimised so if this gets too slow may
need to be re-implmened but his should allow experimnts.