Went through mstyle section: Some wording tweaks
Added statement that a group of mstyle attrs are not part of core (not in a table, so otherwise not marked)

Changed example from one that doesn't work in core to one that does work. The example is probably too long, but if I'm giving a continued fraction, I wanted to use one that is interesting.