Alignment examples
Change the examples in the alto use fake MathML with tables as suggested by @davidcarlisle in #354

The first example in #table-cells-that-are-not-divided-into-alignment-groups is not right though. The long first cell is spacing out the rest of the cells. I played with `table-layout` but didn't see a solution there. Maybe something with width and overflow might work. @davidcarlisle: do you know how to get this right?