Tabular (#357)
* Tabular math fixes

Remove allowed unitless value mention and then add mention that it was allowed in MathML 3 but deprecated in 4 (two places)

A few wording tweaks

Remove some images/<pre>s from the examples.

I only did a cursory review of the Alignment Markers section since that may be deprecated.

* missing </p>

* Alignment examples

Change the examples in the alto use fake MathML with tables as suggested by @davidcarlisle in #354

The first example in #table-cells-that-are-not-divided-into-alignment-groups is not right though. The long first cell is spacing out the rest of the cells. I played with `table-layout` but didn't see a solution there. Maybe something with width and overflow might work. @davidcarlisle: do you know how to get this right?

* fix indentation of long attribute

* html validity, add tbody and delete spurious /span

* overflow table heading DUPLICATE, remove before merging

* remove extra example and center text in an example

Co-authored-by: David Carlisle <d.p.carlisle@gmail.com>