The main change is to make another attempt at clarifying when a core or open (if also part of the AT's dictionary) is a match.
I did _not_ generalize "match" to include the looser concept of matches caused by training some NMT (neural machine translation). The concepts are the same, although obviously NMTs do not do dictionary lookups -- the training data would essentially be the result of a dictionary lookup.

I also made a number of editorial changes including fixing some spelling errors, and unifying wording.

There is, I think, still a problem that we use "Intent Concept Dictionary" and also use the word "list". This might make sense because we provide core and open _lists_, but the implementation looks things up in a _dictionary_ to see if there is a match. I tried to be consistent in the text about this, but I don't believe the text makes this clear.